+++ libmisc/console.c
@@ -70,7 +70,7 @@ static bool is_listed (const char *cfgin, const char *tty, bool def)
 	 * See if this tty is listed in the console file.
 	 */
 
-	while (fgets (buf, (int) sizeof (buf), fp) != NULL) {
+	while (fgets (buf, sizeof (buf), fp) != NULL) {
 		/* Remove optional trailing '\n'. */
 		buf[strcspn (buf, "\n")] = '\0';
 		if (strcmp (buf, tty) == 0) {
