+++ libmisc/ttytype.c
@@ -43,7 +43,7 @@ void ttytype (const char *line)
 		perror (typefile);
 		return;
 	}
-	while (fgets (buf, (int) sizeof buf, fp) == buf) {
+	while (fgets (buf, sizeof buf, fp) == buf) {
 		if (buf[0] == '#') {
 			continue;
 		}
